mdns name fixed
mdns name needs to be unique, similar to the service name.
USE_MARVEL_SWITCH symbol changed to 1. This effectively comments out the reset_phy code. This is because the method of communicating with the phy is not implemented in hardware so could potentially cause errors (although it didn't seem to). Now we simply assume 100MB full connection (it always is) and move on.
This commit is used to create 1.1.22 (REV002 PCB) and 1.1.23 (REV004 production PCB) FW releases and all are loaded onto VUnet server.
First major test with speakers with this release.